(feat) only do pyproject and pip if there is a pipfile

+++ pythonsh/python.sh
@@ -824,10 +824,14 @@ SHELL
         $EDITOR python.sh || exit 1
         git add python.sh
 
-        echo -n ">>>regenerating pyproject.toml."
+        # only regen if there is a pipfile, some projects aren't python
+        if [[ -f Pipfile ]]
+        then
+          echo -n ">>>regenerating pyproject.toml."
 
-        $0 project
-        git add pyproject.toml
+          $0 project
+          git add pyproject.toml
+        fi
       fi
 
       if [[ -z $resume || $resume == "merge" ]]
@@ -850,19 +854,22 @@ SHELL
 
       if [[ -z $resume ||  $resume == "merge" || $resume == "pipfile" ]]
       then
-        test -d releases || mkdir releases
-        test -f Pipfile && pipenv lock
+        if [[ -f Pipfile ]]
+        then
+          test -d releases || mkdir releases
+          test -f Pipfile && pipenv lock
 
-        git add Pipfile.lock
+          git add Pipfile.lock
 
-        VER_PIP="releases/Pipfile-$VERSION"
-        VER_LOCK="releases/Pipfile.lock-$VERSION"
+          VER_PIP="releases/Pipfile-$VERSION"
+          VER_LOCK="releases/Pipfile.lock-$VERSION"
 
-        test -f Pipfile.lock && cp Pipfile.lock $VER_LOCK
-        test -f Pipfile && cp Pipfile $VER_PIP
+          test -f Pipfile.lock && cp Pipfile.lock $VER_LOCK
+          test -f Pipfile && cp Pipfile $VER_PIP
 
-        test -f $VER_PIP && git add $VER_PIP
-        test -f $VER_LOCK && git add $VER_LOCK
+          test -f $VER_PIP && git add $VER_PIP
+          test -f $VER_LOCK && git add $VER_LOCK
+        fi
       fi
 
       if [[ -z $resume || $resume == "merge" || $resume == "pipfile" || $resume == "commit" ]]
